requirements.txt
aiohttp==3.8.4
aiolimiter==1.0.0
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
isort==5.12.0
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.4
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
PyMySQL==1.0.3
python-dotenv==0.21.0
python-telegram-bot==20.2
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.9
tiktoken==0.3.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
sqlalchemy_seeder==0.3.0
